move stuff into components and static respectively
