+++ front-end/src/components/Footer.css
@@ -1,0 +1,11 @@
+.App-footer {
+    background-color: #187c4e;
+    min-height: 15vh;
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    justify-content: center;
+    font-size: calc(10px + 2vmin);
+    color: white;
+}
+
