+++ front-end/src/components/Header.css
@@ -1,0 +1,21 @@
+.Header {
+    text-align: center;
+    background-color: #187c4e;
+  }
+  
+  .logo {
+    height: 12vmin;
+    border: 0.5vmin solid #fda500;
+  }
+  
+  .App-header {
+    background-color: #187c4e;
+    min-height: 15vh;
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    justify-content: center;
+    font-size: calc(10px + 2vmin);
+    color: white;
+  }
+
