+++ front-end/src/index.js
@@ -2,8 +2,8 @@ import React from 'react';
 import ReactDOM from 'react-dom/client';
 import './index.css';
 import App from './App';
-import Header from './Header'
-import Footer from './Footer';
+import Header from './components/Header'
+import Footer from './components/Footer';
 import reportWebVitals from './reportWebVitals';
